# services/travelogue-users/travelogue-users/src/main/java/com/travelogue/users/api/LoginApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.2.3).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.travelogue.users.api;

import javax.validation.Valid;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import com.travelogue.users.models.Error;
import com.travelogue.users.models.Login;
import com.travelogue.users.models.User;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
import io.swagger.annotations.Authorization;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2017-08-18T18:11:14.373Z")

@Api(value = "login", description = "the login API")
public interface LoginApi {

    @ApiOperation(value = "Login User", notes = "Login User API", response = User.class, tags={ "Login User", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Success", response = User.class),
        @ApiResponse(code = 400, message = "Bad Request", response = Error.class),
        @ApiResponse(code = 401, message = "Authentication Failed", response = Void.class),
        @ApiResponse(code = 403, message = "Forbidden (Account Locked)", response = Void.class),
        @ApiResponse(code = 500, message = "Server Vault", response = Void.class) })
    
    @RequestMapping(value = "/login",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<User> loginUser(@ApiParam(value = ""  )  @Valid @RequestBody Login login);


    @ApiOperation(value = "Logout User", notes = "Logout User API", response = Void.class, authorizations = {
        @Authorization(value = "JWT")
    }, tags={ "Logout User", })
    @ApiResponses(value = { 
        @ApiResponse(code = 204, message = "Success", response = Void.class),
        @ApiResponse(code = 401, message = "Authentication Failed", response = Void.class),
        @ApiResponse(code = 500, message = "Server Vault", response = Void.class) })
    
    @RequestMapping(value = "/login",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.DELETE)
    ResponseEntity<Void> logoutUser();

}